applications/kafka-consumer/requirements.txt
# applications/kafka-consumer/requirements.txt
boto3>=1.26.0
kafka-python>=2.0.2
psycopg2-binary>=2.9.0
tabulate>=0.9.0
colorama>=0.4.6